src/third_party/blink/web_tests/external/wpt/service-workers/service-worker/resource-timing-cross-origin.https-expected.txt

This is a testharness.js-based test.
FAIL Test that timing allow check fails when service worker changes origin from same to cross origin. assert_unreached: unexpected rejection: assert_equals: expected 1 but got 0 Reached unreachable code
Harness: the test ran to completion.
